requirements.txt
fastapi==0.109.2
uvicorn[standard]>=0.22.0
tortoise-orm>=0.20.0
aerich>=0.7.0
redis>=4.5.0
celery>=5.2.0
jinja2>=3.1.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
boto3>=1.26.0
httpx>=0.24.0
email-validator>=2.0.0
requests
PyYAML
python-jose[cryptography]>=3.3.0
asyncmy>=0.2.5
cryptography>=3.4.0
tomlkit>=0.11.0
